Readme and Refactor Added common Readme for FE and BE Refacted code like renamign folders and spellings
