+++ .gitignore
@@ -1,1 +1,31 @@
-client/package-lock.json
+**/package-lock.json
+**/node_modules
+**.vscode
+*.log
+*.idea/
+# JetBrains IDE
+.idea/
+
+# Unit test reports
+TEST*.xml
+
+# Generated by MacOS
+.DS_Store
+
+# Generated by Windows
+Thumbs.db
+
+# Applications
+*.app
+*.exe
+*.war
+
+# Large media files
+*.mp4
+*.tiff
+*.avi
+*.flv
+*.mov
+*.wmv
+
+.eslintcache
